Threats Data Views: introduce controlled columns
