+++ projects/js-packages/components/components/threats-data-views/constants.ts
@@ -30,7 +30,6 @@ export const THREAT_ICONS = {
 	default: shieldIcon,
 };
 
-export const THREAT_FIELD_THREAT = 'threat';
 export const THREAT_FIELD_TITLE = 'title';
 export const THREAT_FIELD_DESCRIPTION = 'description';
 export const THREAT_FIELD_ICON = 'icon';
@@ -44,6 +43,40 @@ export const THREAT_FIELD_SIGNATURE = 'signature';
 export const THREAT_FIELD_FIRST_DETECTED = 'first-detected';
 export const THREAT_FIELD_FIXED_ON = 'fixed-on';
 export const THREAT_FIELD_AUTO_FIX = 'auto-fix';
+
+/**
+ * DataViews fields relevant to vulnerable extension threats.
+ * Example: Threat results generated by the free Protect Report feature.
+ */
+export const VULNERABILITY_FIELDS = [
+	THREAT_FIELD_TITLE,
+	THREAT_FIELD_DESCRIPTION,
+	THREAT_FIELD_ICON,
+	THREAT_FIELD_TYPE,
+	THREAT_FIELD_EXTENSION,
+	THREAT_FIELD_PLUGIN,
+	THREAT_FIELD_THEME,
+];
+
+/**
+ * DataViews fields related to threats.
+ * Example: Threat results detected by Jetpack Scan.
+ */
+export const THREAT_FIELDS = [
+	THREAT_FIELD_TITLE,
+	THREAT_FIELD_DESCRIPTION,
+	THREAT_FIELD_ICON,
+	THREAT_FIELD_STATUS,
+	THREAT_FIELD_TYPE,
+	THREAT_FIELD_EXTENSION,
+	THREAT_FIELD_PLUGIN,
+	THREAT_FIELD_THEME,
+	THREAT_FIELD_SEVERITY,
+	THREAT_FIELD_SIGNATURE,
+	THREAT_FIELD_FIRST_DETECTED,
+	THREAT_FIELD_FIXED_ON,
+	THREAT_FIELD_AUTO_FIX,
+];
 
 export const THREAT_ACTION_FIX = 'fix';
 export const THREAT_ACTION_IGNORE = 'ignore';
